debian: Replace old plymouth theme package
Even though both packages are now built in the PPA, `apt` still wants to remove `elementary-default-settings` to install `plymouth-theme-elementary`.

I feel like we probably didn't need `Breaks` or `Replaces` at all because the package name is still the same, we've just moved it to a different source package. Local builds of this work correctly on my machine.